Add PythonSolver derive
